requirements.txt
streamlit
pillow
diffusers